start part 2

+++ tags/models.py
@@ -6,7 +6,7 @@ from django.contrib.contenttypes.fields import GenericForeignKey
 class TaggedItemManager(models.Manager):
     def get_tags_for(self, obj_type, obj_id):
         content_type = ContentType.objects.get_for_model(obj_type)
-        
+
         return TaggedItem.objects \
             .select_related('tag') \
             .filter(
@@ -14,19 +14,17 @@ class TaggedItemManager(models.Manager):
                 object_id=obj_id
             )
 
-# Create your models here.
+
 class Tag(models.Model):
     label = models.CharField(max_length=255)
 
     def __str__(self) -> str:
         return self.label
 
+
 class TaggedItem(models.Model):
     objects = TaggedItemManager()
-    #What tag applied to what object
     tag = models.ForeignKey(Tag, on_delete=models.CASCADE)
-    # Type
-    # ID
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveBigIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
